requirements.txt
sphinxcontrib-issuetracker
tqdm
requests
pytest
pytest-cov
codecov
vcrpy
PyYAML
